Introduce `ChatMessageActionsRouter`
