+++ Sources_v3/StreamChatUI/ChatChannel/ChatMessageListRouter.swift
@@ -24,28 +24,6 @@ open class ChatMessageListRouter<ExtraData: ExtraDataTypes>: ChatRouter<ChatMess
 
         root.present(popup, animated: true)
     }
-
-    open func showMessageDeletionConfirmationAlert(confirmed: @escaping (Bool) -> Void) {
-        let alert = UIAlertController(
-            title: L10n.Message.Actions.Delete.confirmationTitle,
-            message: L10n.Message.Actions.Delete.confirmationMessage,
-            preferredStyle: .alert
-        )
-
-        alert.addAction(UIAlertAction(
-            title: L10n.Alert.Actions.cancel,
-            style: .cancel,
-            handler: { _ in confirmed(false) }
-        ))
-        alert.addAction(UIAlertAction(
-            title: L10n.Alert.Actions.delete,
-            style: .destructive,
-            handler: { _ in confirmed(true) }
-        ))
-
-        let presenter = rootViewController?.presentedViewController ?? rootViewController
-        presenter?.present(alert, animated: true)
-    }
     
     open func showPreview(for attachment: _ChatMessageAttachment<ExtraData>) {
         let preview = ChatAttachmentPreviewVC()
